+++ src/scripts/modules/ex-dropbox/actions/ApplicationActions.js
@@ -1,9 +1,39 @@
-//import actions from '../../components/InstalledComponentsActionCreators';
-import oAuthActions from '../../components/OAuthActionCreators';
-//import store from '../../components/stores/InstalledComponentsStore';
+export function getBucketsForSelection(buckets) {
+  return buckets.map((bucketName) => {
+    return { value: bucketName, label: bucketName };
+  });
+}
 
-const COMPONENT_ID = 'ex-dropbox';
+export function filterBuckets(buckets) {
+  let inputBuckets = buckets.filter((bucket) => {
+    return bucket.get('stage') === 'in';
+  });
 
-export function deleteCredentials(configId) {
-  oAuthActions.deleteCredentials(COMPONENT_ID, configId);
+  return inputBuckets;
+}
+
+export function listBucketNames(buckets) {
+  var bucketNameList = [];
+
+  buckets.map((bucketObject, bucketId) => {
+    bucketNameList.push(bucketId);
+  });
+
+  return bucketNameList;
+}
+
+export function getDestinationName(fileName) {
+  let destinationFile = fileName.toString().replace(/\//g, '_').toLowerCase().slice(1, -4);
+
+  return `${destinationFile}`;
+}
+
+export function extractValues(inputArray) {
+  var returnArray = [];
+
+  inputArray.map((value) => {
+    returnArray.push(value.label);
+  });
+
+  return returnArray;
 }
